Double-linked list
A double-linked list is a data structure in which elements, known as nodes, are linked together in a linear sequence. Each node contains a data element and two pointers, one pointing to the previous node (prev) and another pointing to the next node (next). The first node in the list is called the head, and the last node is called the tail.

Here's a summary of a basic implementation of a double-linked list in Python: